backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
spacy==3.6.1
PyPDF2==3.0.1
python-docx==0.8.11
sentence-transformers==2.2.2
joblib==1.3.1
Werkzeug==2.3.7
